Create a hook for the local list details.
