+++ netflix-create-react-vite-app/src/hooks/useLocalListDetails.ts
@@ -1,0 +1,110 @@
+import { useState, useEffect } from 'react';
+import { VITE_API_KEY } from '../utils/api';
+import { MovieResult } from '../utils/types/types';
+import { useMyList, MyListItem } from '../context/myListContext';
+
+interface UseLocalListDetailsReturn {
+  localMovies: MovieResult[];
+  localLoading: boolean;
+  localError: string | null;
+  failedItems: any[];
+  removalNotice: string | null;
+}
+
+export const useLocalListDetails = (): UseLocalListDetailsReturn => {
+  const { myList, removeFromList } = useMyList();
+  const [localMovies, setLocalMovies] = useState<MovieResult[]>([]);
+  const [localLoading, setLocalLoading] = useState<boolean>(false);
+  const [localError, setLocalError] = useState<string | null>(null);
+  const [failedItems, setFailedItems] = useState<any[]>([]);
+  const [removalNotice, setRemovalNotice] = useState<string | null>(null);
+
+  useEffect(() => {
+    if (!myList || myList.length === 0) {
+      setLocalMovies([]);
+      setFailedItems([]);
+      return;
+    }
+
+    // Filter for valid items only
+    const validList = myList.filter(
+      (item) =>
+        ['movie', 'tv'].includes(item.media_type) && typeof item.id === 'number'
+    );
+
+    if (validList.length === 0) {
+      setLocalMovies([]);
+      setFailedItems(myList); // All items are invalid
+      return;
+    }
+
+    const abortController = new AbortController();
+
+    const fetchLocalMovies = async () => {
+      setLocalLoading(true);
+      setLocalError(null);
+      setFailedItems([]); // Reset failed items on each fetch
+
+      try {
+        const movies: MovieResult[] = [];
+        const failed: any[] = [];
+
+        for (const item of validList) {
+          const url = `https://api.themoviedb.org/3/${item.media_type}/${item.id}?api_key=${VITE_API_KEY}`;
+          const res = await fetch(url, { signal: abortController.signal });
+
+          if (!res.ok) {
+            console.warn(
+              'Failed to fetch:',
+              url,
+              'Status:',
+              res.status,
+              'Item:',
+              item
+            );
+            failed.push(item);
+            removeFromList(item);
+            setRemovalNotice(
+              `"${
+                item.title || item.name || 'Unknown Title'
+              }" was removed from your list because it could not be loaded.`
+            );
+            setTimeout(() => setRemovalNotice(null), 4000);
+            continue;
+          }
+
+          const data = await res.json();
+          if (item.media_type === 'tv') {
+            data.title = data.name;
+          }
+          movies.push(data);
+        }
+
+        setLocalMovies(movies);
+        setFailedItems(failed);
+      } catch (err) {
+        if (!abortController.signal.aborted) {
+          setLocalError('Failed to fetch your list. Please try again later.');
+        }
+      } finally {
+        if (!abortController.signal.aborted) {
+          setLocalLoading(false);
+        }
+      }
+    };
+
+    fetchLocalMovies();
+
+    return () => {
+      abortController.abort();
+    };
+  }, [myList, removeFromList]);
+
+  return {
+    localMovies,
+    localLoading,
+    localError,
+    failedItems,
+    removalNotice,
+  };
+};
